Handle music ids consistently as int Fix many compiler warnings of implicit conversion loses integer precision: 'long' to 'int'. Add helper methods to fetch int result from query

+++ xbmc/filesystem/MusicDatabaseDirectory/QueryParams.cpp
@@ -24,7 +24,7 @@ CQueryParams::CQueryParams()
 
 void CQueryParams::SetQueryParam(NODE_TYPE NodeType, const std::string& strNodeName)
 {
-  long idDb=atol(strNodeName.c_str());
+  int idDb = atoi(strNodeName.c_str());
 
   switch (NodeType)
   {
