+++ xbmc/music/Artist.h
@@ -32,7 +32,7 @@ public:
 class CArtist
 {
 public:
-  long idArtist = -1;
+  int idArtist = -1;
   bool operator<(const CArtist& a) const
   {
     if (strMusicBrainzArtistID.empty() && a.strMusicBrainzArtistID.empty())
@@ -164,7 +164,7 @@ public:
   void SetScrapedMBID(bool scrapedMBID) { this->m_bScrapedMBID = scrapedMBID; }
 
 private:
-  long idArtist = -1;
+  int idArtist = -1;
   std::string m_strArtist;
   std::string m_strSortName;
   std::string m_strMusicBrainzArtistID;
@@ -176,7 +176,7 @@ typedef std::vector<CArtistCredit> VECARTISTCREDITS;
 
 const std::string BLANKARTIST_FAKEMUSICBRAINZID = "Artist Tag Missing";
 const std::string BLANKARTIST_NAME = "[Missing Tag]";
-const long BLANKARTIST_ID = 1;
+const int BLANKARTIST_ID = 1;
 const std::string VARIOUSARTISTS_MBID = "89ad4ac3-39f7-470e-963a-56509c546377";
 
 #define ROLE_ARTIST 1  //Default role
@@ -189,7 +189,7 @@ public:
     : idRole(-1), m_strRole(std::move(strRole)), m_strArtist(std::move(strArtist)), idArtist(-1)
   {
   }
-  CMusicRole(int role, std::string strRole, std::string strArtist, long ArtistId)
+  CMusicRole(int role, std::string strRole, std::string strArtist, int ArtistId)
     : idRole(role),
       m_strRole(std::move(strRole)),
       m_strArtist(std::move(strArtist)),
@@ -199,8 +199,8 @@ public:
   std::string GetArtist() const { return m_strArtist; }
   std::string GetRoleDesc() const { return m_strRole; }
   int GetRoleId() const { return idRole; }
-  long GetArtistId() const { return idArtist; }
-  void SetArtistId(long iArtistId) { idArtist = iArtistId;  }
+  int GetArtistId() const { return idArtist; }
+  void SetArtistId(int iArtistId) { idArtist = iArtistId;  }
 
   bool operator==(const CMusicRole& a) const
   {
@@ -213,7 +213,7 @@ private:
   int idRole;
   std::string m_strRole;
   std::string m_strArtist;
-  long idArtist;
+  int idArtist;
 };
 
 typedef std::vector<CMusicRole> VECMUSICROLES;
